JUPYTER NOTEBOOK EXAMPLE AND VARIABLE BATCH
I added a jupyter notebook to have a more visual perception of the
network results. The notebook is the aplication of the "main.py".
I also fixed a broken import for the lastest tensorflow
version (1.4.0).
I also commented the "multi_dimensional_rnn_while_loop" function
to make it a little more clear (at least IMO) and added variable
batch support.
Finally added the tensorboard command to be able to explore the
graph.